Add products.ts, getPrimaryAppProductId, update imports
