+++ packages/components/src/internal/app/constants.ts
@@ -11,13 +11,12 @@ import { SAMPLE_MANAGER_SEARCH_PLACEHOLDER, SEARCH_PLACEHOLDER } from '../compon
 import { GENERAL_ASSAY_PROVIDER_NAME } from '../components/assay/constants';
 
 import { AppProperties } from './models';
-
-// These ids should match what is used by the MenuProviders in the Java code, so we can avoid toLowerCase comparisons.
-export const LKS_PRODUCT_ID = 'LabKeyServer';
-const BIOLOGICS_PRODUCT_ID = 'Biologics';
-const LIMS_PRODUCT_ID = 'LIMS';
-const SAMPLE_MANAGER_PRODUCT_ID = 'SampleManager';
-const FREEZER_MANAGER_PRODUCT_ID = 'FreezerManager';
+import {
+    BIOLOGICS_PRODUCT_ID,
+    FREEZER_MANAGER_PRODUCT_ID,
+    LIMS_PRODUCT_ID,
+    SAMPLE_MANAGER_PRODUCT_ID,
+} from './products';
 
 const SAMPLE_MANAGER_PRODUCT_NAME = 'Sample Manager';
 const BIOLOGICS_PRODUCT_NAME = 'Biologics';
