+++ packages/components/src/internal/components/productnavigation/utils.ts
@@ -1,6 +1,6 @@
 import { User } from '../base/models/User';
-import { isLKSSupportEnabled, resolveModuleContext } from '../../app/utils';
-import { ModuleContext } from '../base/ServerContext';
+import { isLKSSupportEnabled } from '../../app/utils';
+import { ModuleContext, resolveModuleContext } from '../base/ServerContext';
 
 /**
  * Returns true for the LKB app or for other apps w/ premium module when the user isAdmin or the Look and Feel Setting
